kill old decoding logic
