--- src/debug/decoding/calldata/decode.ts
@@ -1,431 +1,0 @@
-import { Address, bytesToUtf8 } from "@ethereumjs/util";
-import {
-    AddressType,
-    ArrayType,
-    assert,
-    BoolType,
-    BytesType,
-    FixedBytesType,
-    InferType,
-    IntType,
-    PointerType,
-    DataLocation as SolDataLocation,
-    StringType,
-    StructDefinition,
-    TupleType,
-    TypeNode,
-    UserDefinedType
-} from "solc-typed-ast";
-import {
-    bigEndianBufToBigint,
-    fits,
-    MAX_ARR_DECODE_LIMIT,
-    readMem,
-    uint256
-} from "../../../utils/misc";
-import { changeToLocation, isABITypeStaticSized } from "../../../utils/solidity";
-import { CalldataLocation, DataLocation, DataLocationKind, Memory } from "../../types";
-
-export function cd_readMem(
-    loc: CalldataLocation,
-    len: Uint8Array | bigint | number,
-    calldata: Memory
-): Uint8Array | undefined {
-    return readMem(loc.address + loc.base, len, calldata);
-}
-
-export function cd_decodeInt(
-    typ: IntType,
-    loc: CalldataLocation,
-    calldata: Uint8Array
-): undefined | [bigint, number] {
-    const bytes = cd_readMem(loc, 32n, calldata);
-
-    // OoB access
-    if (bytes === undefined) {
-        return undefined;
-    }
-
-    let res = bigEndianBufToBigint(bytes);
-
-    // Convert signed negative 2's complement values
-    if (typ.signed && (res & (BigInt(1) << BigInt(typ.nBits - 1))) !== BigInt(0)) {
-        // Mask out any 1's above the number's size
-        res = res & ((BigInt(1) << BigInt(typ.nBits)) - BigInt(1));
-        res = -((BigInt(1) << BigInt(typ.nBits)) - res);
-    }
-
-    assert(
-        fits(res, typ),
-        `Decoded value ${res} from ${loc} doesn't fit in expected type ${typ.pp()}`
-    );
-
-    return [res, 32];
-}
-
-function cd_decodeAddress(loc: CalldataLocation, calldata: Memory): undefined | [Address, number] {
-    const bytes = cd_readMem(loc, 32, calldata);
-
-    if (bytes === undefined) {
-        return undefined;
-    }
-
-    const res = new Address(bytes.slice(12));
-    return [res, 32];
-}
-
-function cd_decodeFixedBytes(
-    typ: FixedBytesType,
-    loc: CalldataLocation,
-    calldata: Memory
-): undefined | [Uint8Array, number] {
-    const res = cd_readMem(loc, typ.size, calldata);
-    return res === undefined ? undefined : [res, 32];
-}
-
-function cd_decodeBool(loc: CalldataLocation, calldata: Memory): undefined | [boolean, number] {
-    const bytes = cd_readMem(loc, 32, calldata); // @todo is 32 correct here?
-
-    if (bytes === undefined) {
-        return undefined;
-    }
-
-    const res = bigEndianBufToBigint(bytes) !== BigInt(0);
-
-    return [res, 32];
-}
-
-function cd_decodeBytes(loc: CalldataLocation, calldata: Memory): undefined | [Uint8Array, number] {
-    let bytesOffset = loc.address;
-    let bytesSize = 0;
-
-    const len = cd_decodeInt(uint256, loc, calldata);
-
-    if (len == undefined || len[0] >= MAX_ARR_DECODE_LIMIT) {
-        return undefined;
-    }
-
-    const numLen = Number(len[0]);
-
-    bytesOffset += BigInt(len[1]);
-    bytesSize += len[1];
-    bytesSize += numLen + (numLen % 32 === 0 ? 0 : 1 - (numLen % 32));
-
-    const res = cd_readMem(
-        { kind: loc.kind, address: bytesOffset, base: loc.base },
-        numLen,
-        calldata
-    );
-
-    return res === undefined ? undefined : [res, bytesSize];
-}
-
-function cd_decodeString(loc: CalldataLocation, calldata: Memory): undefined | [string, number] {
-    const bytes = cd_decodeBytes(loc, calldata);
-
-    if (bytes === undefined) {
-        return undefined;
-    }
-
-    const str = bytesToUtf8(bytes[0]);
-
-    return [str, bytes[1]];
-}
-
-export function cd_decodeArrayContents(
-    abiType: ArrayType,
-    origType: ArrayType | undefined,
-    arrBaseOffset: bigint,
-    numLen: number,
-    calldata: Memory,
-    infer: InferType
-): undefined | [any[], number] {
-    const res: any[] = [];
-    const elT = origType !== undefined ? origType.elementT : undefined;
-
-    let arrBytesSize = 0;
-    let elOffset = 0n;
-
-    for (let i = 0; i < numLen; i++) {
-        const elementTuple = cd_decodeValue(
-            abiType.elementT,
-            elT,
-            { kind: DataLocationKind.CallData, address: elOffset, base: arrBaseOffset },
-            calldata,
-            infer
-        );
-
-        if (elementTuple === undefined) {
-            return undefined;
-        }
-
-        const [elementVal, elementSize] = elementTuple;
-
-        res.push(elementVal);
-
-        elOffset += BigInt(elementSize);
-        arrBytesSize += elementSize;
-    }
-
-    return [res, arrBytesSize];
-}
-
-function cd_decodeArray(
-    abiType: ArrayType,
-    origType: ArrayType | undefined,
-    loc: CalldataLocation,
-    calldata: Memory,
-    infer: InferType
-): undefined | [any[], number] {
-    let arrOffset = loc.address + loc.base;
-    let arrBytesSize = 0;
-
-    let len: [bigint, number] | undefined;
-
-    if (abiType.size !== undefined) {
-        len = [abiType.size, 0];
-    } else {
-        len = cd_decodeInt(uint256, loc, calldata);
-    }
-
-    if (len == undefined || len[0] >= MAX_ARR_DECODE_LIMIT) {
-        return undefined;
-    }
-
-    const numLen = Number(len[0]);
-
-    arrOffset += BigInt(len[1]);
-    arrBytesSize += len[1];
-
-    const contentsRes = cd_decodeArrayContents(
-        abiType,
-        origType,
-        arrOffset,
-        numLen,
-        calldata,
-        infer
-    );
-
-    if (contentsRes === undefined) {
-        return undefined;
-    }
-
-    return [contentsRes[0], arrBytesSize + contentsRes[1]];
-}
-
-function cd_decodeTuple(
-    abiType: TupleType,
-    origType: TypeNode | undefined,
-    loc: CalldataLocation,
-    calldata: Memory,
-    infer: InferType
-): undefined | [any[], number] {
-    const tupleRes: any[] = [];
-
-    let origElementTs: TypeNode | TypeNode[] | undefined = undefined;
-
-    if (origType instanceof PointerType) {
-        origType = origType.to;
-    }
-
-    if (origType instanceof ArrayType) {
-        assert(
-            origType.size !== undefined && Number(origType.size) === abiType.elements.length,
-            `Expected original type to be a fixed size array of size ${abiType.elements.length} not {0}`,
-            origType
-        );
-
-        origElementTs = origType.elementT;
-    } else if (origType instanceof UserDefinedType) {
-        const def = origType.definition;
-
-        assert(
-            def instanceof StructDefinition && def.vMembers.length === abiType.elements.length,
-            `Expected struct with ${abiType.elements.length} fields not {0}`,
-            origType
-        );
-
-        try {
-            origElementTs = def.vMembers.map((fieldDef) =>
-                changeToLocation(
-                    infer.variableDeclarationToTypeNode(fieldDef),
-                    SolDataLocation.CallData
-                )
-            );
-        } catch (e) {
-            return undefined;
-        }
-    } else if (origType instanceof TupleType) {
-        const origNonNullTypes = origType.elements.filter((t) => t !== null) as TypeNode[];
-        assert(
-            abiType.elements.length === origNonNullTypes.length,
-            `Unexpected mismatch between abi tuple {0} and original tuple {1} in decoding`,
-            abiType,
-            origType
-        );
-        origElementTs = origNonNullTypes;
-    } else if (origType !== undefined) {
-        throw new Error(
-            `Unexpected original type ${origType.pp()} for abi tuple type ${abiType.pp()}`
-        );
-    }
-
-    let tupleOffset: bigint = 0n;
-    const tupleBase = loc.address + loc.base;
-    let size = 0;
-
-    for (let i = 0; i < abiType.elements.length; i++) {
-        const fieldT = abiType.elements[i];
-
-        const origElementT =
-            origElementTs === undefined
-                ? undefined
-                : origElementTs instanceof TypeNode
-                  ? origElementTs
-                  : origElementTs[i];
-
-        assert(fieldT !== null, ``);
-
-        const decodeRes = cd_decodeValue(
-            fieldT,
-            origElementT,
-            { kind: loc.kind, address: tupleOffset, base: tupleBase },
-            calldata,
-            infer
-        );
-
-        if (decodeRes === undefined) {
-            return undefined;
-        }
-
-        const [elVal, elementSize] = decodeRes;
-
-        tupleOffset += BigInt(elementSize);
-        size += elementSize;
-
-        tupleRes.push(elVal);
-    }
-
-    if (origType === undefined || origType instanceof ArrayType || origType instanceof TupleType) {
-        return [tupleRes, size];
-    }
-
-    const fields = (origType.definition as StructDefinition).vMembers;
-    const structRes: any = {};
-
-    for (let i = 0; i < fields.length; i++) {
-        structRes[fields[i].name] = tupleRes[i];
-    }
-
-    return [structRes, size];
-}
-
-function cd_decodePointer(
-    abiType: PointerType,
-    origType: PointerType | undefined,
-    loc: CalldataLocation,
-    calldata: Memory,
-    infer: InferType
-): undefined | [any, number] {
-    const offRes = cd_decodeInt(uint256, loc, calldata);
-
-    if (offRes === undefined) {
-        return undefined;
-    }
-
-    // Adjust relative pointers read **FROM** calldata by the current base offset
-    const size = offRes[1];
-
-    const pointedToLoc: DataLocation = {
-        kind: DataLocationKind.CallData,
-        address: offRes[0],
-        base: loc.base
-    };
-
-    const origPointedToType = origType === undefined ? undefined : origType.to;
-
-    const pointedToValue = cd_decodeValue(
-        abiType.to,
-        origPointedToType,
-        pointedToLoc,
-        calldata,
-        infer
-    );
-
-    if (pointedToValue === undefined) {
-        return undefined;
-    }
-
-    return [pointedToValue[0], size];
-}
-
-export function cd_decodeValue(
-    abiType: TypeNode,
-    origType: TypeNode | undefined,
-    loc: CalldataLocation,
-    calldata: Memory,
-    infer: InferType
-): undefined | [any, number] {
-    /*
-    console.error(
-        `cd_decodeValue(${abiType.pp()}, ${origType ? origType.pp() : undefined}, ${ppLoc(loc)})`
-    );
-    */
-
-    if (abiType instanceof IntType) {
-        return cd_decodeInt(abiType, loc, calldata);
-    }
-
-    if (abiType instanceof AddressType) {
-        return cd_decodeAddress(loc, calldata);
-    }
-
-    if (abiType instanceof FixedBytesType) {
-        return cd_decodeFixedBytes(abiType, loc, calldata);
-    }
-
-    if (abiType instanceof BoolType) {
-        return cd_decodeBool(loc, calldata);
-    }
-
-    if (abiType instanceof ArrayType) {
-        assert(
-            origType === undefined || origType instanceof ArrayType,
-            `Unexpected original type {0}`,
-            origType
-        );
-
-        return cd_decodeArray(abiType, origType, loc, calldata, infer);
-    }
-
-    if (abiType instanceof BytesType) {
-        return cd_decodeBytes(loc, calldata);
-    }
-
-    if (abiType instanceof StringType) {
-        return cd_decodeString(loc, calldata);
-    }
-
-    if (abiType instanceof TupleType) {
-        return cd_decodeTuple(abiType, origType, loc, calldata, infer);
-    }
-
-    if (abiType instanceof PointerType) {
-        assert(
-            origType === undefined || origType instanceof PointerType,
-            `Unexpected original type {0}`,
-            origType
-        );
-
-        // Fixed-sized data types are emitted inline, not in the dynamic region
-        if (isABITypeStaticSized(abiType)) {
-            const origPointedToType = origType === undefined ? undefined : origType.to;
-
-            return cd_decodeValue(abiType.to, origPointedToType, loc, calldata, infer);
-        }
-
-        return cd_decodePointer(abiType, origType, loc, calldata, infer);
-    }
-
-    throw new Error(`NYI decoding ${abiType.pp()}`);
-}
