+++ src/debug/decoding/calldata/index.ts
@@ -1,2 +1,1 @@
-export { cd_decodeArrayContents, cd_decodeValue } from "./decode";
 export { makeCalldataViews } from "./view";
