+++ src/debug/decoding/memory/index.ts
@@ -1,2 +1,1 @@
-export { mem_decodeValue } from "./decoding";
 export { makeMemoryView } from "./view";
