+++ src/debug/decoding/stack/index.ts
@@ -1,2 +1,1 @@
-export * from "./stack";
 export { makeStackView } from "./view";
