+++ src/debug/decoding/storage/index.ts
@@ -1,2 +1,1 @@
-export { nextWord, roundLocToType, stor_decodeMap, stor_decodeValue } from "./decoding";
 export { makeStorageView } from "./view";
